profiler uses fixed viewport
