+++ editor/assets/effects/editor/terrain-select-brush.effect.meta
@@ -1,5 +1,5 @@
 {
-  "ver": "1.4.6",
+  "ver": "1.4.7",
   "importer": "effect",
   "imported": true,
   "uuid": "9b4d3fa5-89bf-4715-b69a-dbbd1efaf6a5",
